+++ agent/src/main/java/com/walmartlabs/concord/agent/ExecutionException.java
@@ -23,10 +23,6 @@ package com.walmartlabs.concord.agent;
 
 public class ExecutionException extends Exception {
 
-    public ExecutionException(String message) {
-        super(message);
-    }
-
     public ExecutionException(String message, Throwable cause) {
         super(message, cause);
     }