+++ agent/src/main/java/com/walmartlabs/concord/agent/JobExecutorException.java
@@ -20,7 +20,6 @@ package com.walmartlabs.concord.agent;
  * =====
  */
 
-
 public class JobExecutorException extends RuntimeException {
 
     public JobExecutorException(String message) {